azure-functions/requirements.txt
azure-functions
azure-storage-blob
azure-keyvault-secrets
redis
celery
firebase-admin
pandas
yfinance
prophet
plotly
requests
python-dotenv 